+++ src/imxrt1060/mod.rs
@@ -4,6 +4,7 @@
 //! functions are tagged with `imxrt-iomuxc` traits.
 
 mod adc;
+mod flexcan;
 mod flexpwm;
 mod lpi2c;
 mod lpspi;